Added Chronic.strip_tokens
This method takes the same options as parse(), it just does the inverse
of what Chronic.parse does - rather than return the parsed time message
this method returns the rest of the string, those tokens that aren't
tagged as time-related.

The rationale behind this is to enable Chronic to perform in quick
calendaring additions (a la Google Calendar's Quick Add feature).